Rename project from carve to shardate
