+++ docker-compose.yml
@@ -1,7 +1,7 @@
 services:
   dev:
     build: .
-    container_name: carve
+    container_name: shardate
     working_dir: /app
     volumes:
       - .:/app
