+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "carve"
+name = "shardate"
 version = "0.1.0"
 description = "A lightweight Python library for efficiently reading year-month-day partitioned Parquet datasets."
 readme = "README.md"
@@ -38,4 +38,4 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/carve"]
+packages = ["src/shardate"]
