+++ src/shardate/__init__.py
@@ -1,6 +1,6 @@
 """Carve: Fast reader for YMD-partitioned Parquet files."""
 
-from carve.read import read_between, read_by_date, read_by_dates
+from shardate.read import read_between, read_by_date, read_by_dates
 
 all = [
     "read_between",
