+++ tests/test_dates.py
@@ -1,5 +1,5 @@
 from datetime import date
-from carve.dates import all_dates_between, is_end_of_month
+from shardate.dates import all_dates_between, is_end_of_month
 
 
 def test_all_dates_between():
